Small shooter changes

+++ src/main/java/frc/robot/Constants.java
@@ -57,7 +57,7 @@ public final class Constants
     public final static int SHOOTER_PIVOTTWO_CAN = 17;
     public final static int SHOOTER_PIVOTTHREE_CAN = 12;
     public final static double FALL_CANCEL_SPEED = 0.1;
-    public final static double PIVOT_CURRENT_LIMIT = 10;
+    public final static int PIVOT_CURRENT_LIMIT = 10;
     public final static int PIVOT_CANCODER_ID = 21;
     //pid controller
     public final static double SHOOTER_KP = 0.001;
